+++ src/opensearch/connection.py
@@ -164,7 +164,9 @@ class BufferedAsyncHttpConnection(AsyncHttpConnection):
             if callable(self._http_auth):
                 req_headers = {
                     **req_headers,
-                    **self._http_auth(method, url, query_string, body),
+                    **self._http_auth(
+                        method=method, url=url, body=body, headers=req_headers
+                    ),
                 }
 
             start = self.loop.time()